split apart test.rs
