+++ tests/checks/mod.rs
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: Apache-2.0
+
+mod unused;
+mod validate;
