+++ tests/feedthroughs/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: Apache-2.0
+
+mod feedthrough;
+mod funnel;
+mod pipeline;
