+++ tests/interfaces/mod.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+
+mod instantiate;
+mod intf;
+mod port;
+mod port_slice;
