+++ tests/syntax/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+mod enums;
+mod import;
+mod inout;
+mod params;
+mod protected;
